Improved read of metadata (#143)
